requirements.txt
Flask
flask-socketio
gevent
Pillow
numpy
requests